+++ src/components/ebay-pagination/index.marko
@@ -64,8 +64,6 @@ static {
             ...state.nextItem.htmlAttributes
             aria-disabled=(state.nextItem.disabled && "true")
             aria-label=(input.a11yNextText || 'Next page')
-
-
             class=state.nextItem.class
             style=state.nextItem.style
             href=state.nextItem.href
