+++ src/components/ebay-pill/browser.json
@@ -5,8 +5,6 @@
                 "@ebay/skin/pill"
             ]
         },
-        "../ebay-button",
-        "require: ./index.js",
-        "./style.less"
+        "../ebay-button"
     ]
 }
