+++ src/components/ebay-pill/index.marko
@@ -1,0 +1,57 @@
+import processHtmlAttributes from "../../common/html-attributes"
+
+static {
+    var ignoredAttributes = [
+        "class",
+        "style",
+        "disabled",
+        "href",
+        "priority",
+        "pressed",
+        "ariaSelectedText"
+    ];
+}
+
+$ var classPrefix = input.href ? "fake-pill" : "pill";
+<span ...processHtmlAttributes(input, ignoredAttributes) class="pill__wrapper">
+    <${input.href ? "a" : "button"}
+        onClick("handleButtonClick")
+        class=[
+            classPrefix,
+            state.pressed
+                ? classPrefix + "--selected"
+                : classPrefix + "--unselected",
+            input.class
+        ]
+        style=input.style
+        href=input.href
+        disabled=input.disabled
+        type=(!input.href && "button")
+        aria-pressed=(!input.href && state.pressed && "true")
+        key="pill"
+        id:scoped="pill">
+        <if(state.text)>
+            <if(state.text.length > 42)>
+                <span aria-hidden="true">${state.text.slice(0, 39)}...</span>
+                <span class="clipped">${state.text}</span>
+            </if>
+            <else>${state.text}</else>
+        </if>
+        <else>
+            <span key="pill-raw-text" id:scoped="pill-raw-text">
+                <if(typeof input.renderBody === "function")>
+                    <${input}/>
+                </if>
+                <else>${input.renderBody}</else>
+            </span>
+        </else>
+        <if(input.href && state.pressed)>
+            <span
+                class="pill__active-text clipped"
+                key="active-text"
+                id:scoped="active-text">
+                - ${input.a11yActiveText || "Selected"}
+            </span>
+        </if>
+    </>
+</span>
