--- src/components/ebay-pill/template.marko
@@ -1,34 +1,0 @@
-<script marko-init>
-    var processHtmlAttributes = require("../../common/html-attributes");
-    var ignoredAttributes = ["class", "style", "disabled", "href", "priority", "pressed", "ariaSelectedText"];
-</script>
-
-<var classPrefix=(data.href ? "fake-pill" : "pill")/>
-
-<span w-bind class="pill__wrapper" ${processHtmlAttributes(data, ignoredAttributes)}>
-    <${data.href ? "a" : "button"}
-        w-id="pill"
-        w-on-click="handleButtonClick"
-        class=[
-            classPrefix,
-            data.pressed ? classPrefix + "--selected" : classPrefix + "--unselected",
-            data.class
-        ]
-        style=data.style
-        href=data.href
-        disabled=data.disabled
-        type=(!data.href && "button")
-        aria-pressed=(!data.href && data.pressed && "true")>
-        <if(data.text)>
-            <if(data.text.length > 42)>
-                <span aria-hidden="true">${data.text.slice(0,39)}...</span>
-                <span class="clipped">${data.text}</span>
-            </if>
-            <else>${data.text}</else>
-        </if>
-        <span else w-id="pill-raw-text" w-body/>
-        <if(data.href && data.pressed)>
-            <span w-id="active-text" class="pill__active-text clipped"> - ${data.a11yActiveText || "Selected"}</span>
-        </if>
-    </>
-</span>
